feat: Mason Tool installer
This tool allows me to install not only lsps, but formatters, linters and debuggers.
It also has a really neat feature, that allows you to only install in certain conditions...
This was such a massive issuer for me in the past as I needed to install a bunch of tools
extra to get everything to work.